ref: rename winners entity to winners list
