+++ src/entities/index.ts
@@ -1,3 +1,3 @@
 export * from './base-repository';
 export * from './user';
-export * from './winners';
+export * from './winners-list';
