+++ src/entities/winners-list.ts
